- Fix ReadProcessMemory and Toolhelp32ReadProcessMemory definitions. - Numerous 64bit fixes. - Remove useless cast. - Remove Fls API exports as per trunk.
svn path=/branches/ros-amd64-bringup/; revision=36475

+++ reactos/dll/win32/kernel32/misc/utils.c
@@ -318,7 +318,7 @@ STDCALL
 BasepFreeStack(HANDLE hProcess,
                PINITIAL_TEB InitialTeb)
 {
-    ULONG Dummy = 0;
+    SIZE_T Dummy = 0;
     
     /* Free the Stack */
     NtFreeVirtualMemory(hProcess,
